\name{SimData}
\alias{SimData}
\docType{data}
\title{Simulated Dataset}
\description{
These data were simulated as described in Section 2.4 of Lock et al., 2013. There are two simulated sources, with rank 1 joint structure and rank 1 structure individual to each source.      
}
\usage{data(SimData)}
\format{
 This dataset is a list of two entries:
\itemize{
\item Data[[1]] (Data$Data1): 50 variables (rows) and 100 samples (columns)
\item Data[[2]] (Data$Data2): 50 variables (rows) and 100 samples (columns)
}
The 100 columns are shared by the sources.}
   
\references{
Lock, EF, KA Hoadley, JS Marron, and AB Nobel. 2013. ''Joint and Individual Variation Explained (JIVE) for Integrated Analysis of Multiple Data Types.'' The Annals of Applied Statistics 7 (1): 523-42.

O'Connell, M. J., & Lock, E.F. (2016). R.JIVE for Exploration of Multi-Source Molecular Data. Bioinformatics advance access: 10.1093/bioinformatics/btw324.
}

\keyword{datasets}
