\name{SimResults}
\alias{Results}
\docType{data}
\title{JIVE Results for Simulated Dataset}
\description{
JIVE results for the simulated data SimData, which were simulated as described in Section 2.4 of Lock et al., 2013. There are two simulated sources, with rank 1 joint structure and rank 1 structure individual to each source.  These results are obtained by running JIVE with permutation testing to select the ranks, and other defaults.        
}
\usage{data(SimResults)}
\format{
 Results: an object of class 'jive'.}
   
\references{
Lock, EF, KA Hoadley, JS Marron, and AB Nobel. 2013. ''Joint and Individual Variation Explained (JIVE) for Integrated Analysis of Multiple Data Types.'' The Annals of Applied Statistics 7 (1): 523-42.

O'Connell, M. J., & Lock, E.F. (2016). R.JIVE for Exploration of Multi-Source Molecular Data. Bioinformatics advance access: 10.1093/bioinformatics/btw324.
}

\keyword{datasets}
